Sync Candle backend, support MXFP4/MXFP6, CUDA 12.9 (#1566)
* Use cudarc 16 PR

* Use float8 0.3.0

* Range fixes

* Handle f4 and others in uqff

* Remove matmul alpha beta

* Some fixes

* More fixes

* More fixes

* Add unfold, mul and act

* Fixes for slice assign

* Fix warn

* Updated deps

* Updated deps

* Updated deps

* Updated deps

* Updated deps

* Updated paged attn

* Handle mistralrs-quant ops

* Handle mistralrs-quant rotary

* Handle hqq

* Handle gptq

* Handle bitsandbytes

* Handle blockwise fp8

* Handle marlin gptq

* Fix some cublaslt errors

* Fix memory usage

* Fix memory usage

* Update argsort

* Update argsort

* Update cublaslt matmul.rs

* Update cublaslt api

* Cuda clippy

* Metal clippy

* Support loading the new dtypes

* Fix flash attn

* Updated candle

* Updated candle

* Updated candle

* Add feature flag set by cuda version

+++ mistralrs-core/src/paged_attention/mod.rs
@@ -141,7 +141,7 @@ pub fn calculate_cache_config(
     // Cap Metal GPU memory to the wired (non‑paged) allocation limit reported by the kernel (`iogpu.wired_limit_mb`).
     // Users can raise this limit with `sudo sysctl -w iogpu.wired_limit_mb=<desired_mb>`.
     let mem_gpu = if matches!(device, Device::Metal(_)) {
-        let metal_cap_mb = MemoryUsage.get_total_memory(device)? as usize / SIZE_IN_MB;
+        let metal_cap_mb = MemoryUsage.get_total_memory(device)? / SIZE_IN_MB;
 
         info!("Metal GPU wired limit is {metal_cap_mb} MB.");
